pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "annotis"
version = "0.1.0"
description = "Domain-agnostic local image annotation and ML-dataset preparation tool"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "PyQt6>=6.4.0,<6.5.0",
    "Pillow>=10.2.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "tifffile>=2024.1.1",
    "piexif>=1.1.3",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
]

[project.scripts]
annotis = "annotis.ui.app:main"

[project.optional-dependencies]
sam = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
]
dev = [
    "black>=24.3.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "pytest-qt>=4.4.0",
    "factory-boy>=3.3.0",
    "pre-commit>=3.7.0",
    "pip-audit>=2.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/annotis"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM", "S"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/annotis/ui/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "raise NotImplementedError",
]
